services/notification-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
pika==1.3.2
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
asyncio-mqtt==0.16.1
jinja2==3.1.2
email-validator==2.1.0
loguru==0.7.2
setuptools==80.9.0
